Update Getting Started Guide
Mainly adjusted text so as to make it clear that this is the Flixel
Community version of the library, and the Flixel Community version of
Flixel is recommended, but not required.

Also changed text to be less "version specific" (so that no major
adjustments are needed in the future), as well as changes as a result of
a modified folder structure.